requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.21.0
webdriver-manager==4.0.1
RestrictedPython>=6.0
PyMuPDF>=1.23.0

